Changes to combat classes in engine and added class Knight in game project.

+++ PixEngine/Scene/Objects/DamageDealer.h
@@ -17,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
     float damageStrength;
 }
 
-- (id) initWithHealth:(float)hp damageStrength:(float)theDamageStrength;
+- (id) initWithHealth:(int)hp damageStrength:(float)theDamageStrength;
 
 @end
